ta: apply __no_stack_protector to __ta_entry() or __ta_entry_c()
Apply the __no_stack_protector attribute to the first C function called
following the TA entry point (i.e., __ta_entry(), or for the special
case of ARM32, __ta_entry_c()). This is required because
__stack_chk_guard is initialized in this very function, therefore stack
protection cannot be assumed to be functional at this point.

Fixes a TA crash on QEMUv7 with Clang 18.1.7 [1].

Link: https://github.com/OP-TEE/optee_os/issues/7047 [1]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Khoa Hoang <admin@khoahoang.com>