mk: add global-cppflags-y
Add global-cppflags-y as a new sub.mk variable. The content of the
variable is added to the C preprocessor flags passed to all files
compiled in the current submodule, $(sm), similarly to how the content
of global-incdirs-y is passed on.

This flag is intended for libraries that in addition to exporting an
include path also need certain defines set.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>